done fragmentation handlers and rework tests
