--- sh/categories/get/la.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-#
-curl -v '127.0.0.1:4221/categories?paginate=%7B"offset"%3A1%2C"limit"%3A7%7D'
