+++ sh/news/get/filterDataAtKnownUser.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 #
-curl -v 'login1:qpass1@127.0.0.1:4221/news?filter=%5B%7B"contents"%3A"2024-08-01"%2C"tag"%3A"FilterDataAt"%7D%5D'
+curl -v 'login1:qpass1@127.0.0.1:4221/news?filter=%5B%7B"contents"%3A"2025-03-16"%2C"tag"%3A"FilterDataAt"%7D%5D'
