+++ src/Database/Authorization.hs
@@ -25,10 +25,10 @@ validCopyRight connString login title = do
         fetchUserFromNews :: (MonadIO m) => SqlPersistT m [Value T.Text]
         fetchUserFromNews = select $ do
           (news :& user) <-
-            from
-              $ table @News
+            from $
+              table @News
                 `innerJoin` table @User
-              `on` (\(n :& u) -> n ^. NewsUserId ==. (u ^. UserId))
+                  `on` (\(n :& u) -> n ^. NewsUserId ==. (u ^. UserId))
           where_ (news ^. NewsTitle ==. (val . getTitle) title)
           pure (user ^. UserLogin)
 
@@ -45,9 +45,9 @@ validPassword connString login password = do
     fetchSaltAndPassword :: (MonadFail m, MonadIO m) => SqlPersistT m [Value T.Text]
     fetchSaltAndPassword = select $ do
       (user :& pass) <-
-        from
-          $ table @User
+        from $
+          table @User
             `innerJoin` table @Password
-          `on` (\(u :& p) -> u ^. UserPasswordId ==. p ^. PasswordId)
+              `on` (\(u :& p) -> u ^. UserPasswordId ==. p ^. PasswordId)
       where_ (user ^. UserLogin ==. (val . getLogin) login)
       pure (pass ^. PasswordQuasiPassword)
