+++ src/Handlers/Database/User.hs
@@ -1,9 +1,9 @@
 module Handlers.Database.User (Handle (..)) where
 
-import Handlers.Database.Base (Offset (..), Limit(..), Success(..))
 import Control.Exception (SomeException)
 import qualified Data.Text as T
 import Data.Time (UTCTime)
+import Handlers.Database.Base (Limit (..), Offset (..), Success (..))
 import qualified Handlers.Logger
 import Handlers.Web.User.Types (UserInternal (..))
 import Schema (User (..))
@@ -12,8 +12,7 @@ import Types (Login (..), PasswordUser (..))
 type HashPasswordUser = T.Text
 
 data Handle m = Handle
-  { 
-    logger :: Handlers.Logger.Handle m,
+  { logger :: Handlers.Logger.Handle m,
     userOffset :: Int,
     userLimit :: Int,
     getTime :: m UTCTime,
@@ -21,4 +20,4 @@ data Handle m = Handle
     pullAllUsers :: Offset -> Limit -> m (Either SomeException [User]),
     findUserByLogin :: Login -> m (Either SomeException (Maybe User)),
     putUser :: UserInternal -> UTCTime -> m (Either SomeException Success)
-  }        
+  }
