+++ src/Handlers/Database/User/Get.hs
@@ -4,8 +4,7 @@ import Control.Exception (displayException)
 import Control.Monad (when)
 import Data.Either (isLeft)
 import qualified Data.Text as T
-import Handlers.Database.Base ( Limit (..), Offset (..))
--- import Handlers.Web.Base (HandleUser (..))
+import Handlers.Database.Base (Limit (..), Offset (..))
 import Handlers.Database.User (Handle (..))
 import Handlers.Logger (Log (..), logMessage)
 import Schema (User (..))
