+++ src/Handlers/Web/Category.hs
@@ -1,4 +1,3 @@
--- {-# LANGUAGE DuplicateRecordFields #-}
 {-# LANGUAGE DataKinds #-}
 
 module Handlers.Web.Category (Handle (..)) where
@@ -6,14 +5,12 @@ module Handlers.Web.Category (Handle (..)) where
 import Data.Binary.Builder (Builder)
 import qualified Data.ByteString as B
 import Data.Text (Text)
+import qualified Handlers.Database.Category
 import qualified Handlers.Logger
 import Network.Wai (Request, Response)
-import qualified Handlers.Database.Category
-
 
 data Handle m = Handle
-  { 
-    logger :: Handlers.Logger.Handle m,
+  { logger :: Handlers.Logger.Handle m,
     base :: Handlers.Database.Category.Handle m,
     getBody :: Request -> m B.ByteString,
     response200 :: Response,
