+++ test/Handlers/Database/Category/CreateSpec.hs
@@ -2,11 +2,11 @@ module Handlers.Database.Category.CreateSpec where
 
 import Control.Monad.State (State, execState, gets, modify)
 import Database.Data.FillTables (cat1, cat2, cat3, cat4, cat5, cat6, cat7, cat8, cat9)
-import Handlers.Database.Base ( Success (..))
+import Handlers.Database.Base (Success (..))
 import Handlers.Database.Category (Handle (..))
 import Handlers.Database.Category.Create (createCategoryBase)
-import Handlers.Web.Base (CategoryInternal (..))
 import qualified Handlers.Logger
+import Handlers.Web.Base (CategoryInternal (..))
 import Schema (Category (..))
 import Test.Hspec
 import Types (Label (..))
@@ -33,13 +33,13 @@ spec = do
                       then Just (Category label undefined)
                       else Nothing,
             Handlers.Database.Category.putCategory =
-               \(CategoryInternal (MkLabel label) _parent) -> do
+              \(CategoryInternal (MkLabel label) _parent) -> do
                 modify (Category label undefined :)
                 pure $ Right Put,
             Handlers.Database.Category.editCategory = \_ _ -> pure $ Right Put,
             Handlers.Database.Category.pullAllCategories = \_ _ -> pure $ Right []
           } ::
-            Handlers.Database.Category.Handle (State [Category])
+          Handlers.Database.Category.Handle (State [Category])
 
   it "Success: category does not exist in the database, category \"parent\" exists in the database" $ do
     length (execState (createCategoryBase baseCategoryHandle (CategoryInternal (MkLabel "NewLabel") (Just $ MkLabel "Man"))) categoriesInBase)
