+++ test/Handlers/Router/Image/GetSpec.hs
@@ -2,23 +2,21 @@
 
 module Handlers.Router.Image.GetSpec (spec) where
 
-import Handlers.Web.Image.Get (existingImages)
 import Control.Monad.State (State, evalState, get)
 import Data.Binary.Builder (fromByteString)
 import Data.ByteString.Base64 as B64
 import qualified Data.Text.Encoding as E
 import Database.Data.FillTables (image1, image2, image3)
-
+import qualified Handlers.Database.Image
+import qualified Handlers.Logger
+import qualified Handlers.Web.Image
+import Handlers.Web.Image.Get (existingImages)
 import Network.HTTP.Types (hContentType, status200)
 import Network.Wai (defaultRequest, queryString, rawPathInfo, responseBuilder)
 import Network.Wai.Internal (Response (..))
 import Schema (Image (..))
 import Test.Hspec (Spec, it, shouldBe, shouldNotBe)
 import Types (NumberImage (..))
-
-import qualified Handlers.Web.Image
-import qualified Handlers.Database.Image
-import qualified Handlers.Logger
 import qualified Web.Utils as WU
 
 spec :: Spec
@@ -30,21 +28,22 @@ spec = do
         Handlers.Logger.Handle
           { Handlers.Logger.levelLogger = Handlers.Logger.Debug,
             Handlers.Logger.writeLog = \_ -> pure ()
-            }
+          }
       baseImageHandle =
         Handlers.Database.Image.Handle
           { Handlers.Database.Image.logger = logHandle,
             Handlers.Database.Image.pullImage = \(MkNumberImage num) -> get >>= pure . Right . Just . flip (!!) (fromIntegral num)
-            }
-      imageHandle = Handlers.Web.Image.Handle {
-          Handlers.Web.Image.logger = logHandle,
-          Handlers.Web.Image.base = baseImageHandle,
-          Handlers.Web.Image.response404 = WU.response404,
-          Handlers.Web.Image.response400 = WU.response400,
-          Handlers.Web.Image.response500 = WU.response500,
-          Handlers.Web.Image.mkGoodResponse = WU.mkGoodResponse,
-          Handlers.Web.Image.mkResponseForImage = testImage
-        } ::
+          }
+      imageHandle =
+        Handlers.Web.Image.Handle
+          { Handlers.Web.Image.logger = logHandle,
+            Handlers.Web.Image.base = baseImageHandle,
+            Handlers.Web.Image.response404 = WU.response404,
+            Handlers.Web.Image.response400 = WU.response400,
+            Handlers.Web.Image.response500 = WU.response500,
+            Handlers.Web.Image.mkGoodResponse = WU.mkGoodResponse,
+            Handlers.Web.Image.mkResponseForImage = testImage
+          } ::
           Handlers.Web.Image.Handle (State [Image])
 
   it "Can get an image by the ID" $ do
@@ -54,7 +53,7 @@ spec = do
     let req'' = req {queryString = [("id", Just "2")]}
     evalState (existingImages imageHandle req'') imagesInBase
       `shouldNotBe` testImage (imagesInBase !! 1)
-    
+
 testImage :: Image -> Response
 testImage (Image header base) = responseBuilder status200 [(hContentType, contentType)] content
   where
