+++ test/Handlers/Router/News/CreateSpec.hs
@@ -5,19 +5,19 @@ module Handlers.Router.News.CreateSpec (spec) where
 import Control.Monad.Identity (Identity, runIdentity)
 import Data.Binary.Builder as BU (Builder)
 import Database.Data.FillTables (cat1, time4, user1test)
-import Handlers.Web.News (Handle (..))
+import qualified Handlers.Database.Auth
+import Handlers.Database.Base (Success (..))
+import qualified Handlers.Database.News
 import qualified Handlers.Logger
 import Handlers.Web.Base (NewsInternal (..))
+import Handlers.Web.News (Handle (..))
+import Handlers.Web.News.Create (createNews)
 import Network.HTTP.Types (status200)
 import Network.Wai (defaultRequest, rawPathInfo, responseBuilder)
 import Network.Wai.Internal (Response (..))
-import Test.Hspec (Spec, it, shouldBe)
 import Schema (ColumnType (..), SortOrder (..))
-import qualified Handlers.Database.Auth
-import qualified Handlers.Database.News
+import Test.Hspec (Spec, it, shouldBe)
 import qualified Web.Utils as WU
-import Handlers.Database.Base ( Success (..))
-import Handlers.Web.News.Create (createNews)
 
 spec :: Spec
 spec = do
@@ -29,20 +29,20 @@ spec = do
         Handlers.Logger.Handle
           { Handlers.Logger.levelLogger = Handlers.Logger.Debug,
             Handlers.Logger.writeLog = \_ -> pure ()
-            }
-      client = 
-              Handlers.Database.Auth.Client
-                { Handlers.Database.Auth.clientAdminToken = Nothing,
-                  Handlers.Database.Auth.clientPublisherToken = Nothing,
-                  Handlers.Database.Auth.author = Nothing
-                }
+          }
+      client =
+        Handlers.Database.Auth.Client
+          { Handlers.Database.Auth.clientAdminToken = Nothing,
+            Handlers.Database.Auth.clientPublisherToken = Nothing,
+            Handlers.Database.Auth.author = Nothing
+          }
       authHandle =
         Handlers.Database.Auth.Handle
           { Handlers.Database.Auth.logger = logHandle,
             Handlers.Database.Auth.findUserByLogin = \_ -> pure $ Right Nothing,
             Handlers.Database.Auth.validPassword = \_ _ -> pure $ Right True,
             Handlers.Database.Auth.client = client,
-            Handlers.Database.Auth.validCopyRight = \_ _ -> pure $ Right True 
+            Handlers.Database.Auth.validCopyRight = \_ _ -> pure $ Right True
           }
       baseNewsHandle =
         Handlers.Database.News.Handle
@@ -58,24 +58,25 @@ spec = do
             Handlers.Database.News.findCategoryByLabel = const (pure . Right $ Just cat1),
             Handlers.Database.News.putNews = \(NewsInternal _title _login _label _content _images _isPublish) _time -> pure $ Right Put,
             Handlers.Database.News.findNewsByTitle = const (pure $ Right Nothing),
-            Handlers.Database.News.pullAllNews = \_ _ _columntype _sortorder _find _filters -> pure $ Right [], 
+            Handlers.Database.News.pullAllNews = \_ _ _columntype _sortorder _find _filters -> pure $ Right [],
             Handlers.Database.News.editNews = \_ _ _ -> pure $ Right Change
           }
 
-      newsHandle = Handlers.Web.News.Handle {
-          Handlers.Web.News.logger = logHandle,
-          Handlers.Web.News.base = baseNewsHandle,
-          Handlers.Web.News.auth = authHandle,
-          Handlers.Web.News.client = client,
-          Handlers.Web.News.response400 = WU.response400,
-          Handlers.Web.News.response500 = WU.response500,
-          Handlers.Web.News.response200 = WU.response200,
-          Handlers.Web.News.response404 = WU.response404,
-          Handlers.Web.News.response403 = WU.response403,
-          Handlers.Web.News.mkGoodResponse = testBuilder,
-          Handlers.Web.News.getBody = const . pure $ bodyReq
-        } ::
-         Handlers.Web.News.Handle Identity
+      newsHandle =
+        Handlers.Web.News.Handle
+          { Handlers.Web.News.logger = logHandle,
+            Handlers.Web.News.base = baseNewsHandle,
+            Handlers.Web.News.auth = authHandle,
+            Handlers.Web.News.client = client,
+            Handlers.Web.News.response400 = WU.response400,
+            Handlers.Web.News.response500 = WU.response500,
+            Handlers.Web.News.response200 = WU.response200,
+            Handlers.Web.News.response404 = WU.response404,
+            Handlers.Web.News.response403 = WU.response403,
+            Handlers.Web.News.mkGoodResponse = testBuilder,
+            Handlers.Web.News.getBody = const . pure $ bodyReq
+          } ::
+          Handlers.Web.News.Handle Identity
 
   it "Can create news" $ do
     runIdentity (createNews (error "Publisher Role") newsHandle req')
@@ -91,7 +92,6 @@ instance Show Response where
 instance Eq Response where
   (==) (ResponseBuilder s h b) (ResponseBuilder s' h' b') = (s == s') && (h == h') && (show b == show b')
   (==) _ _ = undefined
-
 
 testBuilder :: Builder -> Response
 testBuilder = responseBuilder status200 []
