+++ test/Handlers/Router/News/EditSpec.hs
@@ -1,25 +1,25 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module Handlers.Router.News.EditSpec (spec) where
+
 import Control.Monad.State (State, evalState, gets)
 import Data.Binary.Builder as BU (Builder)
-import Schema (ColumnType (..), SortOrder (..))
-import Database.Data.FillTables (cat1, news1, news2, news3, news4, user1test, time4)
+import Database.Data.FillTables (cat1, news1, news2, news3, news4, time4, user1test)
+import qualified Handlers.Database.Auth
+import Handlers.Database.Base (Success (..))
+import qualified Handlers.Database.News
 import qualified Handlers.Logger
 import Handlers.Web.Base (NewsEditInternal (..))
 import qualified Handlers.Web.Base as WB
+import Handlers.Web.News (Handle (..))
+import Handlers.Web.News.Update (updateNews)
 import Network.HTTP.Types (status200)
 import Network.Wai (defaultRequest, rawPathInfo, responseBuilder)
 import Network.Wai.Internal (Response (..))
-import Schema (News (..)) 
+import Schema (ColumnType (..), News (..), SortOrder (..))
 import Test.Hspec (Spec, it, shouldBe, shouldNotBe)
 import Types (Title (..))
-import qualified Handlers.Database.Auth
-import qualified Handlers.Database.News
 import qualified Web.Utils as WU
-import Handlers.Database.Base ( Success (..))
-import Handlers.Web.News.Update (updateNews)
-import Handlers.Web.News (Handle (..))
 
 spec :: Spec
 spec = do
@@ -31,14 +31,14 @@ spec = do
         Handlers.Logger.Handle
           { Handlers.Logger.levelLogger = Handlers.Logger.Debug,
             Handlers.Logger.writeLog = \_ -> pure ()
-            }
+          }
       authHandle =
         Handlers.Database.Auth.Handle
           { Handlers.Database.Auth.logger = logHandle,
             Handlers.Database.Auth.findUserByLogin = \_ -> pure $ Right Nothing,
             Handlers.Database.Auth.validPassword = \_ _ -> pure $ Right True,
             Handlers.Database.Auth.client = Handlers.Database.Auth.Client Nothing Nothing Nothing,
-            Handlers.Database.Auth.validCopyRight = \_ _ -> pure $ Right True 
+            Handlers.Database.Auth.validCopyRight = \_ _ -> pure $ Right True
           }
       baseNewsHandle =
         Handlers.Database.News.Handle
@@ -53,38 +53,40 @@ spec = do
             Handlers.Database.News.filtersNews = [],
             Handlers.Database.News.findCategoryByLabel = const (pure . Right $ Just cat1),
             Handlers.Database.News.putNews = \(WB.NewsInternal _title _login _label _content _images _isPublish) _time -> pure $ Right Put,
-            Handlers.Database.News.findNewsByTitle = 
-             \title -> do
-              titles <- gets (map (MkTitle . newsTitle))
-              pure $
-                Right $
-                  if title `elem` titles
-                    then Just undefined
-                    else Nothing,
-            Handlers.Database.News.pullAllNews = \_ _ _columntype _sortorder _find _filters -> pure $ Right [], 
+            Handlers.Database.News.findNewsByTitle =
+              \title -> do
+                titles <- gets (map (MkTitle . newsTitle))
+                pure $
+                  Right $
+                    if title `elem` titles
+                      then Just undefined
+                      else Nothing,
+            Handlers.Database.News.pullAllNews = \_ _ _columntype _sortorder _find _filters -> pure $ Right [],
             Handlers.Database.News.editNews = \_title _time (NewsEditInternal _mbtitle _mblogin _mblabel _mbcontent _image _mbp) -> pure $ Right Change
           }
 
-      newsHandle = Handlers.Web.News.Handle {
-          Handlers.Web.News.logger = logHandle,
-          Handlers.Web.News.base = baseNewsHandle,
-          Handlers.Web.News.auth = authHandle,
-          Handlers.Web.News.client = Handlers.Database.Auth.Client Nothing Nothing Nothing,
-          Handlers.Web.News.response400 = WU.response400,
-          Handlers.Web.News.response500 = WU.response500,
-          Handlers.Web.News.response200 = WU.response200,
-          Handlers.Web.News.response404 = WU.response404,
-          Handlers.Web.News.response403 = WU.response403,
-          Handlers.Web.News.mkGoodResponse = testBuilder,
-          Handlers.Web.News.getBody = const . pure $ bodyReq
-        } ::
-         Handlers.Web.News.Handle (State [News])
---
+      newsHandle =
+        Handlers.Web.News.Handle
+          { Handlers.Web.News.logger = logHandle,
+            Handlers.Web.News.base = baseNewsHandle,
+            Handlers.Web.News.auth = authHandle,
+            Handlers.Web.News.client = Handlers.Database.Auth.Client Nothing Nothing Nothing,
+            Handlers.Web.News.response400 = WU.response400,
+            Handlers.Web.News.response500 = WU.response500,
+            Handlers.Web.News.response200 = WU.response200,
+            Handlers.Web.News.response404 = WU.response404,
+            Handlers.Web.News.response403 = WU.response403,
+            Handlers.Web.News.mkGoodResponse = testBuilder,
+            Handlers.Web.News.getBody = const . pure $ bodyReq
+          } ::
+          Handlers.Web.News.Handle (State [News])
+  --
   it "Author can edit news" $ do
     let authHandle' = authHandle {Handlers.Database.Auth.validCopyRight = \_login _title -> pure $ Right True}
         newsHandle' =
           newsHandle
-            { Handlers.Web.News.auth = authHandle'}
+            { Handlers.Web.News.auth = authHandle'
+            }
     evalState (updateNews (error "Author") newsHandle' req') newsInBase
       `shouldBe` test200
 
@@ -92,7 +94,8 @@ spec = do
     let authHandle' = authHandle {Handlers.Database.Auth.validCopyRight = \_login _title -> pure $ Right False}
         newsHandle' =
           newsHandle
-            { Handlers.Web.News.auth = authHandle'}
+            { Handlers.Web.News.auth = authHandle'
+            }
     evalState (updateNews (error "No-Author") newsHandle' req') newsInBase
       `shouldNotBe` test200
